Started exporter development, read the rest.
The class BasicFBPrinter was re-written to some degree to save the
project in EAE format. It has only been tried with a single handmade
basic FB, and with this the bezierpoints (EAE thing) isn't fnished. This will corrupt the project files when saving,
so please have a back up if you wan't to try. The original class can be
found in the comment block at the bottom.