Expanded SystemEcoConverter.
There's a major flaw when it comes to modifying the system file in FBME,
which causes applications to be inconsistent from resources. THerefore,
I introduced a new function to the EcoConverter, fixInconsistencies. It
should now fix the mismathcing constant inputs in the interface of any
FB in any application (compared to resource).